feito login

+++ app/src/main/java/com/batsworks/budget/ui/state/login/LoginViewModel.kt
@@ -21,102 +21,102 @@ import kotlinx.coroutines.flow.receiveAsFlow
 import kotlinx.coroutines.launch
 
 class LoginViewModel(
-	private val validateName: ValidateName = ValidateName(),
-	private val validatePhone: ValidatePhone = ValidatePhone(),
-	private val validateEmail: ValidateEmail = ValidateEmail(),
-	private val validateRepeatedPassword: ValidateRepeatedPassword = ValidateRepeatedPassword(),
-	private val validatePassword: ValidatePassword = ValidatePassword(),
-	private val validateTerms: ValidateTerms = ValidateTerms(),
-	private val repository: CustomRepository<UserEntity>? = null,
+    private val validateName: ValidateName = ValidateName(),
+    private val validatePhone: ValidatePhone = ValidatePhone(),
+    private val validateEmail: ValidateEmail = ValidateEmail(),
+    private val validateRepeatedPassword: ValidateRepeatedPassword = ValidateRepeatedPassword(),
+    private val validatePassword: ValidatePassword = ValidatePassword(),
+    private val validateTerms: ValidateTerms = ValidateTerms(),
+    private val repository: CustomRepository<UserEntity>? = null,
 ) : ViewModel() {
 
-	private val TAG = LoginViewModel::class.simpleName!!
-	var state by mutableStateOf(RegistrationFormState())
+    private val TAG = LoginViewModel::class.simpleName!!
+    var state by mutableStateOf(RegistrationFormState())
 
-	private val resourceEventChannel = Channel<Resource<Any>>()
-	val resourceEventFlow = resourceEventChannel.receiveAsFlow()
+    private val resourceEventChannel = Channel<Resource<Any>>()
+    val resourceEventFlow = resourceEventChannel.receiveAsFlow()
 
-	fun onEvent(event: RegistrationFormEvent) {
-		when (event) {
-			is RegistrationFormEvent.NameChanged -> {
-				state = state.copy(nome = event.name)
-			}
+    fun onEvent(event: RegistrationFormEvent) {
+        when (event) {
+            is RegistrationFormEvent.NameChanged -> {
+                state = state.copy(nome = event.name)
+            }
 
-			is RegistrationFormEvent.EmailChanged -> {
-				state = state.copy(email = event.email)
-			}
+            is RegistrationFormEvent.EmailChanged -> {
+                state = state.copy(email = event.email)
+            }
 
-			is RegistrationFormEvent.TelefoneChanged -> state =
-				state.copy(telefone = event.telefone)
+            is RegistrationFormEvent.TelefoneChanged -> state =
+                state.copy(telefone = event.telefone)
 
-			is RegistrationFormEvent.PasswordChanged -> {
-				state = state.copy(password = event.password)
-				submitData()
-			}
+            is RegistrationFormEvent.PasswordChanged -> {
+                state = state.copy(password = event.password)
+                submitData()
+            }
 
-			is RegistrationFormEvent.RepeatedPasswordChanged -> {
-				state = state.copy(repeatedPassword = event.repeatedPassword)
-				submitData()
-			}
+            is RegistrationFormEvent.RepeatedPasswordChanged -> {
+                state = state.copy(repeatedPassword = event.repeatedPassword)
+                submitData()
+            }
 
-			is RegistrationFormEvent.TermsChanged -> {
-				state = state.copy(acceptedTerms = event.accepted)
-				submitData()
-			}
+            is RegistrationFormEvent.TermsChanged -> {
+                state = state.copy(acceptedTerms = event.accepted)
+                submitData()
+            }
 
-			else -> submitData()
-		}
-	}
+            else -> submitData()
+        }
+    }
 
-	private fun submitData() {
-		val nameResult = validateName.execute(state.nome)
-		val phoneResult = validatePhone.execute(state.telefone)
-		val emailResult = validateEmail.execute(state.email)
-		val passwordResult = validatePassword.execute(state.password)
-		val repeatedPasswordResult =
-			validateRepeatedPassword.execute(state.password, state.repeatedPassword)
-		val termsResult = validateTerms.execute(state.acceptedTerms)
+    private fun submitData() {
+        val nameResult = validateName.execute(state.nome)
+        val phoneResult = validatePhone.execute(state.telefone)
+        val emailResult = validateEmail.execute(state.email)
+        val passwordResult = validatePassword.execute(state.password)
+        val repeatedPasswordResult =
+            validateRepeatedPassword.execute(state.password, state.repeatedPassword)
+        val termsResult = validateTerms.execute(state.acceptedTerms)
 
-		val hasError = listOf(
-			nameResult,
-			emailResult,
-			phoneResult,
-			passwordResult,
-			repeatedPasswordResult,
-			termsResult
-		).any { !it.successful }
+        val hasError = listOf(
+            nameResult,
+            emailResult,
+            phoneResult,
+            passwordResult,
+            repeatedPasswordResult,
+            termsResult
+        ).any { !it.successful }
 
-		if (hasError) {
-			state = state.copy(
-				nomeError = nameResult.errorMessage,
-				emailError = emailResult.errorMessage,
-				telefoneError = phoneResult.errorMessage,
-				passwordError = passwordResult.errorMessage,
-				repeatedPasswordErro = repeatedPasswordResult.errorMessage,
-				acceptedTermsError = termsResult.errorMessage
-			)
-			Log.d(TAG, "Foram encontrados erros")
-			return
-		}
-	}
+        if (hasError) {
+            state = state.copy(
+                nomeError = nameResult.errorMessage,
+                emailError = emailResult.errorMessage,
+                telefoneError = phoneResult.errorMessage,
+                passwordError = passwordResult.errorMessage,
+                repeatedPasswordErro = repeatedPasswordResult.errorMessage,
+                acceptedTermsError = termsResult.errorMessage
+            )
+            Log.d(TAG, "Foram encontrados erros")
+            return
+        }
+    }
 
-	fun registerUser() {
-		viewModelScope.launch {
-			if (repository == null) {
-				resourceEventChannel.send(Resource.Failure("repository must not be null"))
-				return@launch
-			}
-			val user = UserEntity(
-				nome = state.nome,
-				email = state.email,
-				phone = state.telefone,
-				password = state.password.toInt()
-			)
-			val querySnapshot = repository.findByParam(Filter.equalTo("email", user.email))
+    fun registerUser() {
+        viewModelScope.launch {
+            if (repository == null) {
+                resourceEventChannel.send(Resource.Failure("repository must not be null"))
+                return@launch
+            }
+            val user = UserEntity(
+                nome = state.nome,
+                email = state.email,
+                phone = state.telefone,
+                password = state.password.toLong()
+            )
+            val querySnapshot = repository.findByParam(Filter.equalTo("email", user.email))
 //			if (!querySnapshot.isEmpty && querySnapshot.size() > 1) {
 //				resourceEventChannel.send(Resource.Failure("Email já cadastrado"))
 //				return@launch
-			}
+        }
 
 //			repository.save(user).addOnCompleteListener { complete ->
 //				if (complete.isCanceled)
@@ -128,6 +128,6 @@ class LoginViewModel(
 //			}
 //			resourceEventChannel.send(Resource.Loading(true))
 //		}
-	}
+    }
 
 }
